+++ cypress/integration/blocks.js
@@ -599,7 +599,6 @@ if (Cypress.env('API') !== 'guillotina') {
 
     it.only('Add Table of Contents block', () => {
       const h2 = 'expectedH2';
-      const h3 = 'expectedH3';
 
       cy.get('.block.inner.text .public-DraftEditor-content')
         .type(` ${h2}`)
@@ -607,27 +606,18 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.get(
         '#page-edit .draftJsToolbar__buttonWrapper__1Dmqh:nth-of-type(5)',
       ).click();
+      cy.get('.block.inner.text .public-DraftEditor-content')
+        .click()
+        .type(' {enter}');
+      cy.get('.ui.basic.icon.button.block-add-button').click();
+      cy.get('.title')
+        .contains('Common')
+        .click();
+      cy.get('.ui.basic.icon.button.toc')
+        .contains('Table of Contents')
+        .click();
       cy.get('#toolbar-save').click();
-      // cy.visit('/my-page/edit');
-      cy.get('.toolbar-actions .edit').click();
-
-      // cy.get('.block.text [contenteditable]').click();
-      // cy.get('button.block-add-button').click();
-      // cy.get('.blocks-chooser .title')
-      //   .contains('Common')
-      //   .click();
-      // cy.get('.ui.buttons .button.toc').click();
-      // cy.get('.ui.drag.block.inner.toc')
-      //   .click()
-      //   .type('{enter}');
-
-      // cy.get('.block.text.selected')
-      //   .click()
-      //   .type(` ${h2}`);
-      // cy.get('.block.text.selected').setSelection(` ${h2}`);
-      // cy.get(
-      //   '#page-edit .draftJsToolbar__buttonWrapper__1Dmqh:nth-of-type(5)',
-      // ).click();
+      cy.get(' a').contains(` ${h2}`);
     });
   });
 }